third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-drawImage-video.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

function jsWrapperClass(node)
{
    // returns the ClassName of node
    if (!node)
        return "[null]";
    var string = Object.prototype.toString.apply(node);

    // string will be of the form [object ClassName]
    return string.substr(8, string.length - 9);
}

function shouldBeType(expression, className)
{
    assert_equals(jsWrapperClass(expression), className);
}

function shouldBeClear(canvas, x, y) {
    // should be transparent black pixels
    var ctx = canvas.getContext('2d');
    d = ctx.getImageData(x, y, 1, 1).data;
    assert_array_equals(d, [0, 0, 0, 0], "Pixel must be clear");
}

function clearContext(ctx, ctx1) {
    ctx.clearRect(0, 0, 500, 500);
    ctx1.clearRect(0, 0, 500, 500);
}

var bitmap;
var video;

var canvas = document.createElement("canvas");
canvas.setAttribute("width", "500");
canvas.setAttribute("height", "500");

var canvas1 = document.createElement("canvas");
canvas1.setAttribute("width", "500");
canvas1.setAttribute("height", "500");

video = document.createElement("video");
video.addEventListener("canplaythrough", videoLoaded, false);
video.src = "../../compositing/resources/video.ogv";

function videoLoaded() {
    promise_test(function() {
        return createImageBitmap(video).then(
            function (image) { checkNoCrop(image); });
    }, 'createImageBitmap from video with no crop rect.');

    promise_test(function() {
        return createImageBitmap(video, 0, 0, 100, 100).then(
            function (image) { checkCrop(image, 0, 0, 100, 100); });
    }, 'createImageBitmap from video with crop rect.');

    promise_test(function() {
        return createImageBitmap(video, 50, 50, 100, 100).then(
            function (image) { checkCrop(image, 50, 50, 100, 100); });
    }, 'createImageBitmap from video with crop rect from right.');

    promise_test(function() {
        return createImageBitmap(video, 100, 100, 100, 100).then(
            function (image) { checkCrop(image, 100, 100, 100, 100); });
    }, 'createImageBitmap from video with crop rect from center.');

    promise_test(function() {
        return createImageBitmap(video, -100, -100, 600, 600).then(
            function (image) { checkOverCrop(image); });
    }, 'createImageBitmap from video with over crop.');

    promise_test(function() {
        return createImageBitmap(video, 100, 100, 500, 500).then(
            function (image) { checkOverCropRight(image); });
    }, 'createImageBitmap from video with over crop from right.');

    promise_test(function() {
        return createImageBitmap(video, 100, 100, -100, -100).then(
            function (image) { checkCrop(image, 0, 0, 100, 100); });
    }, 'createImageBitmap from video with negative size crop rect.');

    promise_test(function() {
        return createImageBitmap(video, 0, 0, Math.pow(10, 6), Math.pow(10, 6))
            .then(
                function (image) { checkNoCrop(image); });
    }, 'createImageBitmap from video with huge crop rect.');

    async_test(t => {
        createImageBitmap(video, -300, -300, 300, 300).then(function() {
            assert_not_reached('Asking for an empty ImageBitmap should be rejected');
            t.done();
        }, function() {
            t.done();
        });
    }, "Asking for an empty ImageBitmap should be rejected (1).");

    async_test(t => {
        createImageBitmap(video, 400, 300, 300, 300).then(function() {
            assert_not_reached('Asking for an empty ImageBitmap should be rejected');
            t.done();
        }, function() {
            t.done();
        });
    }, "Asking for an empty ImageBitmap should be rejected (2).");
}

function compareTwoCanvases(sx, sy, sw, sh, ctx, ctx1)
{
    var data = ctx.getImageData(sx, sy, sw, sh).data;
    var data1 = ctx1.getImageData(sx, sy, sw, sh).data;
    var dataMatched = true;
    for (var i = 0; i < data.length; i++) {
        // data and data1 are strictly the same under software rendering, 
        // but significantly different on GPU.
        if (Math.abs(data[i] - data1[i]) > 80) {
            dataMatched = false;
            break;
        }
    }
    assert_true(dataMatched);
}

function checkNoCrop(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType(bitmap, "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (352, 288)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, 0, 0);
    compareTwoCanvases(0, 0, 352, 288, ctx, ctx1);

    // shrunk to (0, 0), (500, 500)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0, 500, 500);
    ctx1.drawImage(video, 0, 0, 500, 500);
    compareTwoCanvases(0, 0, 500, 500, ctx, ctx1);

    // shrunk to (100, 100), (380, 380)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 100, 100, 380, 380);
    ctx1.drawImage(video, 100, 100, 380, 380);
    compareTwoCanvases(100, 100, 380, 380, ctx, ctx1); 

    // black should be drawn to (100, 100), (200, 200)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 100, 100, 100, 100, 100, 100, 100, 100);
    ctx1.drawImage(video, 100, 100, 100, 100, 100, 100, 100, 100);
    compareTwoCanvases(100, 100, 100, 100, ctx, ctx1);

    // nothing should be drawn
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 400, 300, 200, 200, 100, 100, 100, 100);
    shouldBeClear(canvas, 10, 10);
    shouldBeClear(canvas, 80, 80);
    shouldBeClear(canvas, 150, 150);
    shouldBeClear(canvas, 210, 210);
}

function checkCrop(imageBitmap, sx, sy, sw, sh) {
    bitmap = imageBitmap;
    shouldBeType(bitmap, "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (100, 100)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, sx, sy, sw, sh, 0, 0, sw, sh);
    compareTwoCanvases(0, 0, 100, 100, ctx, ctx1);
}

function checkOverCrop(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType(bitmap, "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (352, 288)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, 0, 0, 352, 288);
    compareTwoCanvases(0, 0, 500, 500, ctx, ctx1);
}

function checkOverCropRight(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType(bitmap, "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // should be drawn to (0, 0), (252, 188)
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    ctx1.drawImage(video, 100, 100, 252, 188, 0, 0, 252, 188);
    compareTwoCanvases(0, 0, 252, 188, ctx, ctx1);
}

function checkEmpty(imageBitmap) {
    bitmap = imageBitmap;
    shouldBeType(bitmap, "ImageBitmap");

    var ctx = canvas.getContext("2d");
    var ctx1 = canvas1.getContext("2d");
    // nothing should be drawn
    clearContext(ctx, ctx1);
    ctx.drawImage(imageBitmap, 0, 0);
    shouldBeClear(canvas, 10, 10);
    shouldBeClear(canvas, 90, 90);
    shouldBeClear(canvas, 110, 110);
    shouldBeClear(canvas, 210, 210);
}

</script>
</body>
</html>
